add comment page

+++ src/main/java/life/ledon/community/dto/QuestionDTO.java
@@ -13,7 +13,7 @@ public class QuestionDTO {
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
+    private Long creator;
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
